reusable status screens
