+++ .gitignore
@@ -43,7 +43,7 @@ app.*.map.json
 !/packages/flutter_tools/test/data/dart_dependencies_test/**/.packages
 
 #apiKeys and config
-#apiKey.dart
+apiKey.dart
 config.dart
 
 #test
