+++ lib/view/error.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+
+class Error extends StatelessWidget {
+  final String errorMessage;
+
+  final Function onRetryPressed;
+
+  const Error({Key key, this.errorMessage, this.onRetryPressed})
+      : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: <Widget>[
+          Text(
+            errorMessage,
+            textAlign: TextAlign.center,
+            style: TextStyle(
+              color: Colors.red,
+              fontSize: 18,
+            ),
+          ),
+          SizedBox(height: 8),
+          RaisedButton(
+            color: Colors.redAccent,
+            child: Text(
+              'Retry',
+              style: TextStyle(),
+            ),
+            onPressed: onRetryPressed,
+          )
+        ],
+      ),
+    );
+  }
+}
