+++ pubspec.yaml
@@ -61,6 +61,10 @@ flutter:
   # "family" key with the font family name, and a "fonts" key with a
   # list giving the asset and other descriptors for the font. For
   # example:
+  fonts:
+    - family: Arvo
+      fonts:
+        - asset: assets/fonts/Arvo-Bold.ttf
   # fonts:
   #   - family: Schyler
   #     fonts:
